Move EC2 network metadata support to its own file
